+++ src/handler/regexp.ts
@@ -1,9 +1,0 @@
-import { handlers } from "../seeker";
-import * as _ from 'lodash';
-
-
-handlers.regexp = {
-    match: {
-        fn: ([str]: [string], [regexp]: [RegExp]) => str.match(regexp),
-    },
-}